# apps/web/public/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BIKIE API",
    "version": "1.0.0",
    "description": "Stable v1 contract under `/api/*`. Flutter and web share this surface. `/api/v2` is reserved for approved breaking changes only (ADR-028). Schemas are incrementally detailed; treat `.docs/API.md` as the narrative source of truth."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (Next.js Route Handlers)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Better Auth HTTP-only session cookie (web)."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Better Auth bearer token (Flutter / mobile). Header: Authorization: Bearer <token>."
      },
      "CronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cron routes: Authorization: Bearer <CRON_SECRET>."
      }
    },
...
  },
  "paths": {
...
    "/api/admin/bikes": {
      "post": {
        "operationId": "post_api_admin_bikes",
        "tags": [
          "admin"
        ],
        "summary": "POST /api/admin/bikes",
        "description": "Source: `apps/web/app/api/admin/bikes/route.ts`. Auth class: `admin`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "admin",
        "x-bikie-source": "apps/web/app/api/admin/bikes/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/admin/partners/{id}": {
      "get": {
        "operationId": "get_api_admin_partners_id_",
        "tags": [
          "admin"
        ],
        "summary": "GET /api/admin/partners/{id}",
        "description": "Source: `apps/web/app/api/admin/partners/[id]/route.ts`. Auth class: `admin`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "admin",
        "x-bikie-source": "apps/web/app/api/admin/partners/[id]/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "patch": {
        "operationId": "patch_api_admin_partners_id_",
        "tags": [
          "admin"
        ],
        "summary": "PATCH /api/admin/partners/{id}",
        "description": "Source: `apps/web/app/api/admin/partners/[id]/route.ts`. Auth class: `admin`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "admin",
        "x-bikie-source": "apps/web/app/api/admin/partners/[id]/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "delete": {
        "operationId": "delete_api_admin_partners_id_",
        "tags": [
          "admin"
        ],
        "summary": "DELETE /api/admin/partners/{id}",
        "description": "Source: `apps/web/app/api/admin/partners/[id]/route.ts`. Auth class: `admin`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "admin",
        "x-bikie-source": "apps/web/app/api/admin/partners/[id]/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/partner/application": {
      "get": {
        "operationId": "get_api_partner_application",
        "tags": [
          "partner"
        ],
        "summary": "GET /api/partner/application",
        "description": "Source: `apps/web/app/api/partner/application/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/application/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/partner/application/submit": {
      "post": {
        "operationId": "post_api_partner_application_submit",
        "tags": [
          "partner"
        ],
        "summary": "POST /api/partner/application/submit",
        "description": "Source: `apps/web/app/api/partner/application/submit/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/application/submit/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/partner/bikes": {
      "get": {
        "operationId": "get_api_partner_bikes",
        "tags": [
          "partner"
        ],
        "summary": "GET /api/partner/bikes",
        "description": "Source: `apps/web/app/api/partner/bikes/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/bikes/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "operationId": "post_api_partner_bikes",
        "tags": [
          "partner"
        ],
        "summary": "POST /api/partner/bikes",
        "description": "Source: `apps/web/app/api/partner/bikes/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/bikes/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/partner/profile": {
      "get": {
        "operationId": "get_api_partner_profile",
        "tags": [
          "partner"
        ],
        "summary": "GET /api/partner/profile",
        "description": "Source: `apps/web/app/api/partner/profile/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/profile/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "put": {
        "operationId": "put_api_partner_profile",
        "tags": [
          "partner"
        ],
        "summary": "PUT /api/partner/profile",
        "description": "Source: `apps/web/app/api/partner/profile/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/profile/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  },
...
}